Salvar e obter
Salvar e obter obj

+++ Rating.java
@@ -5,11 +5,11 @@ import java.io.Serializable;
 public class Rating implements Serializable, Comparable<Rating> {
 
 	private static final long serialVersionUID = 1L;
-	private int usuario_id;
+	private long usuario_id;
 	private long isbn;
 	private int rating;
 
-	public Rating(int usuario_id, long isbn, int rating) {
+	public Rating(long usuario_id, long isbn, int rating) {
 		this.setUsuario_id(usuario_id);
 		this.setIsbn(isbn);
 		this.rating = rating;
@@ -20,11 +20,11 @@ public class Rating implements Serializable, Comparable<Rating> {
 		return this.rating - rating.rating;
 	}
 
-	public int getUsuario_id() {
+	public long getUsuario_id() {
 		return usuario_id;
 	}
 
-	public void setUsuario_id(int usuario_id) {
+	public void setUsuario_id(long usuario_id) {
 		this.usuario_id = usuario_id;
 	}
